feat(markdown): refactor models
