--- src/markdown/lexer.spec.ts
@@ -1,5 +1,0 @@
-describe("Lexer", () => {
-  it("should be true", () => {
-    expect(true).toBe(true)
-  })
-})
