+++ src/markdown/options-parser.ts
@@ -1,5 +1,5 @@
 import { BaseParser } from "./base-parser"
-import { OptionInfo } from "./event-models"
+import { OptionInfo } from "./models/option-info"
 import { isDigit, isLetter } from "./utils"
 
 export class OptionsParser extends BaseParser<OptionInfo[]> {
