+++ src/markdown/schedule-parser.spec.ts
@@ -1,4 +1,6 @@
-import { ScheduleEvent, Range, OptionInfo } from "./event-models"
+import { OptionInfo } from "./models/option-info"
+import { ScheduleEvent } from "./models/schedule-event"
+import { TimeInfo } from "./models/time-info"
 import { ScheduleParser } from "./schedule-parser"
 
 describe("ScheduleParser", () => {
@@ -23,7 +25,7 @@ describe("ScheduleParser", () => {
 
     expect(result).toEqual([
       new ScheduleEvent(
-        new Range("11:00", "15:40"),
+        new TimeInfo("11:00", "15:40"),
         "Title",
         [
           new OptionInfo("m", true),
@@ -34,14 +36,14 @@ describe("ScheduleParser", () => {
         [],
       ),
       new ScheduleEvent(
-        "16:00",
+        new TimeInfo("16:00"),
         "Title 2",
         [new OptionInfo("c", "green")],
         ["description line 1"],
         [],
       ),
       new ScheduleEvent(
-        new Range("18:00", "20:00"),
+        new TimeInfo("18:00", "20:00"),
         "Title 3",
         [],
         [],
