+++ src/markdown/schedule-parser.ts
@@ -1,6 +1,6 @@
 import { BaseParser } from "./base-parser"
-import { ScheduleEvent } from "./event-models"
 import { EventParser } from "./event-parser"
+import { ScheduleEvent } from "./models/schedule-event"
 
 export class ScheduleParser extends BaseParser<ScheduleEvent[]> {
   public parse(): ScheduleEvent[] {
