+++ src/markdown/serializer/time-info.serializer.ts
@@ -1,10 +1,10 @@
-import { TimeInfo } from "../event-models"
+import { TimeInfo } from "../models/time-info"
 import { Serializer } from "./serializer.interface"
 
 export class TimeInfoSerializer implements Serializer<TimeInfo> {
   public serialize(timeInfo: TimeInfo): string {
-    if (typeof timeInfo === "string") {
-      return timeInfo
+    if (!timeInfo.end) {
+      return timeInfo.start
     }
 
     return `${timeInfo.start} - ${timeInfo.end}`
