+++ src/markdown/time-parser.ts
@@ -1,29 +1,33 @@
 import { BaseParser } from "./base-parser"
-import { TimeInfo, Range } from "./event-models"
+import { TimeInfo } from "./models/time-info"
 
 export class TimeInfoParser extends BaseParser<TimeInfo> {
   public parse(): TimeInfo {
     return this.choice<TimeInfo>(
       () => this.parseTimeSpan(),
-      () => this.parseTime(),
+      () => this.parseSingleTime(),
     )
   }
 
-  public parseTime(): string {
-    const hours = this.repeat(2, () => this.digit()).join("")
-    this.matchString(":")
-    const minutes = this.repeat(2, () => this.digit()).join("")
-
-    return `${hours}:${minutes}`
+  public parseSingleTime(): TimeInfo {
+    return new TimeInfo(this.parseTime())
   }
 
-  public parseTimeSpan(): Range<string> {
+  public parseTimeSpan(): TimeInfo {
     const start = this.parseTime()
     this.spaces()
     this.matchString("-")
     this.spaces()
     const end = this.parseTime()
 
-    return new Range(start, end)
+    return new TimeInfo(start, end)
+  }
+
+  private parseTime(): string {
+    const hours = this.repeat(2, () => this.digit()).join("")
+    this.matchString(":")
+    const minutes = this.repeat(2, () => this.digit()).join("")
+
+    return `${hours}:${minutes}`
   }
 }
